Extract owned inventory read surfaces
