+++ src/mtg_source_stack/inventory/service.py
@@ -22,15 +22,10 @@ from .access import (
 )
 from .audit import list_inventory_audit_events
 from .analysis import (
-    OWNED_INVENTORY_PAGE_SORT_DIRECTIONS,
-    OWNED_INVENTORY_PAGE_SORT_KEYS,
     build_duplicate_groups_from_owned_rows,
     export_inventory_csv,
     inventory_health,
     inventory_report,
-    list_owned,
-    list_owned_filtered,
-    list_owned_filtered_page,
     list_price_gaps,
     reconcile_prices,
     render_inventory_csv_export,
@@ -49,6 +44,13 @@ from .decklist_import import import_decklist_text
 from .deck_url_import import import_deck_url
 from .inventories import create_inventory, ensure_default_inventory, list_inventories, list_visible_inventories
 from .inventories import summarize_actor_access
+from .owned_items import (
+    OWNED_INVENTORY_PAGE_SORT_DIRECTIONS,
+    OWNED_INVENTORY_PAGE_SORT_KEYS,
+    list_owned,
+    list_owned_filtered,
+    list_owned_filtered_page,
+)
 from .mutations import (
     add_card,
     add_card_with_connection,
